Add test using `NotificationTrace`
Because:
To experiment with `NotificationTrace`.

Using this method is less flexible then using the polling method as we
need to create each assertions independently. If we want to have another
type of assertion, then we should write a new assertion method in
`NotificationTrace` and most likely a new notification stream class to
handle the different `hasMatched` condition

This commit:
- Added test using notification method